Don't round the master audio time multiplier to an integer
A significant complexity increase for an incremental accuracy boost,
but it's probably more worthwhile at low channel counts.

+++ src/backend.h
@@ -27,7 +27,7 @@ struct ChipMetadata {
 struct RenderJobHandle {
     QString name;
     QString path;
-    int time_multiplier;
+    float time_multiplier;
     QFuture<QString> future;
 };
